Add solution to 7a
This one require a LOT of Rust learnings, hence why it took so long.
Also, I was getting stuck on how to use the smart pointer stuff as well
as a bunch of other compile errors and even some weird inconsistent
answers on subsequent runs which somehow magically fixed itself. This
almost made me want to ragequit Rust @_@

In particular, trying to figure out smart pointers and borrowing/mutable
borrowing was a major PITA when creating the tree structure. I actually
gave up the original attempt to make a "real" tree with parent pointers
and went with the simpler arena approach (which I also learned about for
this problem).

Also, my takeaway is to just compile something and then try to do
whatever rustc suggests. That, or just stick & and * around all the Rc's
or RefCells. They don't sound like good takeaways though.