Add solution to 11b
Note that the solution here exploits a property of the input I saw: each
monkey's divisibility test uses a distinct prime number. Because of
that, this solution (ab)uses modular arithmetic to store modulos of the
worry levels instead of the actual worry level to avoid integer overflow
while still being able to correctly apply the divisibility test for each
item.

It was that, or pull in some bignum/arbitrary precision integer crate,
but I'm aiming to stay within the language's standard library, and I
wasn't about to go implement bignums. Also, it's not often that I get to
flex my number theory knowledge :P

So, maybe this approach is technically cheating (in that you're not
supposed to assume the divisibility tests all use coprime modulos)...
but hey, it works for me.