Add solution to 12a
Half of the time was spent optimizing/pruning search states so that it
doesn't take forever to return an answer. And even then, it takes 9s on
my 3900X which still feels like a long time, but not sure what else I
would try to optimize/prune, or what would be considered a reasonable
run time for this problem. Oh well.