Add solution to 14b
This solution runs a little slower than I'd maybe like it to (2.841s on
my 3900X), and I imagine there's probably some geometric-based
computation that can be done in a much faster way, but eh, this is good
enough for now. Might revisit that idea if I still end up curious after
the rest of Advent of Code.