Add solution to 15b
Like 15a, the input definition is modified to take in the max coordinate
for the search space as the first line, since the example and the actual
problem use different search spaces.

Also, like 15a, I feel like performance is not great (~10.5s on my
3900X), but I'm not too sure how to do a more optimal search off the top
of my head. (Maybe some sort of sensor iteration ordering so that the
sensor radius check loop tends to exit early?)