Add solution to 16b
This one was a doozy for a mix of personal and technical reasons...

For one, I've just been having other things vying for my attention over
Advent of Code, and I physically haven't been feeling that great either
(seasonal stuff).

I'm also temporarily working off of my old laptop now instead of on my
desktop. This means I've massively downgraded my CPU from the 3900X to
an i5-5200U. It also turns out the version of Rust in Debian Bullseye
(current stable) does not support the 2021 edition I was using, nor did
downgrading to the 2018 edition seem to work out of the gate, so I
decided to YOLO upgrade my laptop to Debian Bookworm (current testing)
so I could get a newer version of Rust and continue working on this, and
that ended up taking longer than expected.

Then I had to debug some logic issues in my code (which really involved
adding mounds of eprintlns and poring through the output) with all the
different weird cases.

And after all that, I don't actually know if I should consider solution
or the runtime on the input acceptable. It's in O(min) territory which
is bad, but again, this is on my old laptop. Plus, the code is
horrendously long and awful (498 lines wtf!), and I'm sure there's some
reasonable refactoring that could be done (probably helper function and
lambda-ize some of the procedures across the different cases), but I'm
very much at the point where I don't care as much and just wanna move on
x_x