Add solution to 21b
Ideally I would have wanted to memoize/store the has_human computations
for each part of the tree, but it seemed non-trivial to me as to how to
do that with recursion and within Rust's mutable/non-mutable borrow
rules. I then figured I might as well see if it's performant enough even
with checking has_human for each node each time, which it seems to be.

A potential follow-up is to find a clean and elegant way to turn the
has_human call into a look-up, whether it's precomputing it and storing
it within each tree node or memoizing the call somehow.