Add solution to 22b
This one was pretty painful :(

For one, I'm not super happy with having to basically hard code a bunch
of cube net edge-to-edge relations without taking into account all
possible cube net layouts instead of just the ones in the sample and
input (which was confirmed in
https://www.reddit.com/comments/zsgbe7/_/j17smus/). It still feels like
there's some geometric property or axiom to be used to generalize the
approach here for all possible cube net layouts.

Figuring out and putting in all the different edge-to-edge relations was
also a little tedious. It was difficult enough just staring at the net,
so I even resorted to recreating the nets/cubes on little pieces of
paper that I cut out and folded.

Then, I made a typo that added an extra day of debugging when I had
everything running but got the wrong answer. It was actually a bit
interesting thinking of how to efficiently debug/trace through what was
going on. Obviously I had a bunch of eprintln's to verify that I was
parsing the path correctly and then trace through the edge-to-edge
moves, but with the amount of moves/steps that were being made, that
would take too long, so then I had it print out the edge-to-edge moves
it was making and then filtered those out to just the unique set of
moves. And that's when I realized one of the cases wasn't being used
because I typo'd the guard condition.

So yeah, this one felt not that great to solve TBH :(